pyproject.toml
[project]
name = "tjr103-group-1-project"
version = "0.1.0"
description = ""
authors = [
    {name = "Bevis",email = "add412@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "pandas (>=2.3.3,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "numpy (>=2.3.3,<3.0.0)",
    "google-cloud-storage (>=3.4.1,<4.0.0)",
    "pymysql (>=1.1.2,<2.0.0)",
    "ipykernel (>=7.0.0,<8.0.0)",
    "jupyter (>=1.1.1,<2.0.0)",
    "notebook (>=7.4.7,<8.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "matplotlib (>=3.10.7,<4.0.0)",
    "selenium (>=4.36.0,<5.0.0)",
    "googlemaps (>=4.10.0,<5.0.0)",
    "lxml (>=6.0.2,<7.0.0)",
    "thefuzz (>=0.22.1,<0.23.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "xlrd (>=2.0.2,<3.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "geopandas (>=1.1.1,<2.0.0)",
    "opencc-python-reimplemented (>=0.1.7,<0.2.0)",
    "apache-airflow (>=3.1.1,<4.0.0)"
]


[tool.poetry]
packages = [
    {include = "mods", from = "."}
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
